convert MockUserClient to Swift (#3200)
Merge this as https://github.com/wireapp/wire-ios/pull/3197 was approved.

+++ Wire-iOS Tests/Mocks/MockMessage+Creation.swift
@@ -98,7 +98,7 @@ final class MockMessageFactory: NSObject {
         var userClients: [AnyHashable] = []
 
         for user: Any in mockSystemMessageData.users {
-            if let client = (user as? MockUser)?.feature(withUserClients: UInt(numClients)) {
+            if let client = (user as? MockUser)?.feature(withUserClients: numClients) {
                 userClients.append(contentsOf: client)
             }
         }
